+++ package.json
@@ -49,13 +49,12 @@
     "@types/pino-pretty": "^5.0.0",
     "config": "^3.3.9",
     "js-yaml": "^4.1.0",
+    "moment": "^2.29.4",
     "pino-pretty": "^10.2.3",
     "reflect-metadata": "^0.1.13",
-    "tsyringe": "^4.6.0",
     "yargs": "^17.7.2",
     "zod": "^3.22.4",
-    "zx": "^4.3.0",
-    "moment": "^2.29.4"
+    "zx": "^4.3.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^18.2.0",
