+++ src/common/validations.ts
@@ -1,3 +1,4 @@
+
 import fs from 'fs';
 import moment from 'moment';
 
@@ -17,6 +18,6 @@ export const fileExists = (filePath: string): boolean | void => {
 
 export const isValidDateFormat = (dateString: string): boolean | void => {
     const validDateStringFormat = "yyyy-MM-dd'T'HH:mm:ss";
-    const isValidDateFormat = moment(dateString, moment.ISO_8601, true).isValid(); // true
+    const isValidDateFormat = moment(dateString, moment.ISO_8601, true).isValid();
     return isValidDateFormat ? true : throwError(`Date string must be 'ISO_8601' format: ${validDateStringFormat}, for exmaple: 2023-11-07T12:35:00`);
 }
