+++ src/seed.ts
@@ -1,16 +1,15 @@
 import config from 'config';
 import { $ } from 'zx';
-import { createGeojsonTxtFile, createSeedYaml } from './yaml/seedYaml';
+import { createSeedYamlFile } from './yaml/creator';
 
 // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
 export type SeedOptions = { cache: string, fromZoomLevel: number, toZoomLevel: number, refreshBefore: string, wktFilePath: string, skipUncached: boolean };
 export const executeSeed = async (options: SeedOptions): Promise<void> => {
-  // create the seed.yaml file
-  await createSeedYaml(options);
-
   const mapproxyYamlFilePath = config.get<string>('script.mapproxyYamlFilePath');
   const seedYamlFilePath = config.get<string>('script.seedYamlFilePath');
-  let flags: string[] = [];
+  const flags: string[] = [];
+
+  await createSeedYamlFile(options);
 
   if (options.skipUncached) {
     flags.push('--skip-uncached');
